refactor: Simplificar classes e métodos, utilizando records e ajustando tipos de dados

+++ src/main/java/com/safeway/tech/controllers/ResponsavelController.java
@@ -35,7 +35,7 @@ public class ResponsavelController {
     }
 
     @PutMapping("/{idResponsavel}")
-    public Responsavel alterarResponsavel(@RequestBody Responsavel novoResponsavel,@PathVariable int idResponsavel){
+    public Responsavel alterarResponsavel(@RequestBody Responsavel novoResponsavel,@PathVariable Long idResponsavel){
         return responsavelService.alterarResponsavel(novoResponsavel, idResponsavel);
     }
 }
